requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
Flask-Login==0.6.3
stripe==7.10.0
Flask-Session==0.5.0
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-cloud-logging==3.9.0
gunicorn==21.2.0
cryptography>=42.0.0
Pillow==11.1.0
boto3==1.34.34
Flask-Babel==4.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
paramiko==3.4.0 